migration: standards
* Ignores field 962__ except for standards
* closes https://github.com/CERNDocumentServer/cds-migrator-kit/issues/88
