+++ cds_ils/importer/providers/cds/rules/base.py
@@ -378,36 +378,6 @@ def standard_review(self, key, value):
     return _extensions
 
 
-@model.over("publication_info", "^962__")
-def publication_additional(self, key, value):
-    """Translates additional publication info & other related_records field."""
-    _publication_info = self.get("publication_info", [])
-    _migration = self["_migration"]
-    _related = _migration["related"]
-    empty = not bool(_publication_info)
-    for i, v in enumerate(force_list(value)):
-        temp_info = {}
-        pages = clean_val("k", v, str)
-        if pages:
-            temp_info.update(pages=pages)
-        rel_recid = clean_val("b", v, str)
-        if rel_recid:
-            _related.append(
-                {
-                    "related_recid": rel_recid,
-                    "relation_type": OTHER_RELATION.name,
-                    "relation_description": "is chapter of"
-                }
-            )
-            _migration.update({"related": _related, "has_related": True})
-        if not empty and i < len(_publication_info):
-            _publication_info[i].update(temp_info)
-        else:
-            _publication_info.append(temp_info)
-
-    return _publication_info
-
-
 @model.over("_migration", "(^775__)|(^787__)")
 @out_strip
 def related_records(self, key, value):
